minimal usable state
