+++ src/Fonts.jl
@@ -1,7 +1,7 @@
 module Fonts
 
-using ..Colors
-using ..Plots.Commons
+using Plots.Colors
+using Plots.Commons
 # keep in mind: these will be reexported and are public API
 export font, scalefontsizes, resetfontsizes, text, is_horizontal
